+++ packages/utilities/index.ts
@@ -1,0 +1,24 @@
+
+import Gap from "./src/lib/components/Gap.astro"
+import Projector from "./src/lib/components/Projector.astro"
+
+
+export {
+    iterate,
+     range,
+     iterateRange,
+ } from "./src/lib/helpers"
+
+export type {
+    HasForEachMethod,
+    GetAppropriateFunctionBasedOnWhetherOrNotAGeneratorOfAnIterableWithTheForEachMethodIsPassed,
+     IterationInfo,
+     IterateRangeCallback,
+     IterateRangeOptions
+} from "./src/lib/types"
+
+export {
+    Gap,
+    Projector,
+}
+