+++ packages/flow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forastro/flow
 
+## 0.1.2
+
+### Patch Changes
+
+- Changed where the main file is from src/index to just index types should be from the main package
+- Updated dependencies
+  - @forastro/utilities@1.0.2
+
 ## 0.1.1
 
 ### Patch Changes